worker/requirements.txt
flask==2.2.3
nats-py==2.2.0
flask[async]